[#88] SemesterList UI 리팩토링
